rename views folder
