+++ wls-gui-wahllokalsystem/src/plugins/router.ts
@@ -19,12 +19,12 @@ import { WahlbezirksArtEnum } from "@/types/wahlbezirksArtEnum.ts";
 import EreignisseView from "@/views/EreignisseView.vue";
 import ExampleError404View from "@/views/ExampleError404View.vue";
 import HomeView from "@/views/HomeView.vue";
-import BWBWahlbriefErfassungView from "@/views/wahlvorbereitung/BWBWahlbriefErfassungView.vue";
-import UWBWaehlerverzeichnisView from "@/views/wahlvorbereitung/UWBWaehlerverzeichnisView.vue";
-import UWBWahlhandlungView from "@/views/wahlvorbereitung/UWBWahlhandlungView.vue";
-import WahlbriefZulassungView from "@/views/wahlvorbereitung/WahlbriefZulassungView.vue";
-import WahleroeffnungView from "@/views/wahlvorbereitung/WahleroeffnungView.vue";
-import WahlumgebungView from "@/views/wahlvorbereitung/WahlumgebungView.vue";
+import BWBWahlbriefErfassungView from "@/views/wahlhandlung/BWBWahlbriefErfassungView.vue";
+import UWBWaehlerverzeichnisView from "@/views/wahlhandlung/UWBWaehlerverzeichnisView.vue";
+import UWBWahlhandlungView from "@/views/wahlhandlung/UWBWahlhandlungView.vue";
+import WahlbriefZulassungView from "@/views/wahlhandlung/WahlbriefZulassungView.vue";
+import WahleroeffnungView from "@/views/wahlhandlung/WahleroeffnungView.vue";
+import WahlumgebungView from "@/views/wahlhandlung/WahlumgebungView.vue";
 import WahlvorstandAnwesenheitView from "@/views/WahlvorstandAnwesenheitView.vue";
 
 const permitNavigationOnlyForWahlbezirksArtUwb = () => {
