+++ wls-gui-wahllokalsystem/src/views/wahlhandlung/BWBWahlbriefErfassungView.vue
