+++ wls-gui-wahllokalsystem/src/views/wahlhandlung/WahleroeffnungView.vue
